Cluster.jl Documentation

Overview

This module contains functions to determine natural clusters using unsupervised learning. A point of interest in this module over other libraries is the rich set of metrics that one can use with K-means clustering. Additionally, some of the metrics may be weighted which can be used to help alleviate K-means attraction to spherical clusters.

This module also has the functionality to determine the cluster size of a data set.

Cluster functions (from lowest to highest level):

kmeans_cluster: Given a cluster number and metric, determine clusters.

This clustering algorithm can use any of the metrics: L2 ($L_2$ – default), LP ($L_p$), LI ($L_\infty$), KL (Kullback-Leibler), CD (Cosine Distance), and JD (Jaccard Distance).

In the case of L2 and CD, the metrics allow for weighted distances represented as a positive definite matrix.

find_best_info_for_ks: Given a metric and a range of cluster numbers, determine clusters and gather fitness data for each cluster.

This function uses the functions:

kmeans_cluster.

find_best_cluster: Given a metric, find the best cluster number and determine the clusters.

This function uses the functions:

kmeans_cluster

find_best_info_for_ks

Metric Definitions:

Given N vectors, ${\bf x}, {\bf y}$ :

L2: The standard $L_2$ norm: ${\rm L2}({\bf x}, {\bf y}) = \sqrt{\sum_{i=1}^N (x_i - y_i)^2}$

With a symmetric, positive semi-definite weight matrix W this becomes: ${\rm L2}({\bf x}, {\bf y}, \hbox{M=W}) = \sqrt{{\bf x} {\boldsymbol \cdot} (M {\bf y})}$

LP: The standard $L_p$ norm: ${\rm LP}({\bf x}, {\bf y}, p) = \left(\sum_{i=1}^N |x_i - y_i|^p)\right)^{\frac{1}{p}}$

Note: To use this metric with find_best_cluster for a given value of p, you will need to pass the closure, (x,y; kwargs...) -> LP(x,y, p; kwargs...), to the keyword parameter dmetric.

LI: The standard $L_\infty$ norm: ${\rm LI}({\bf x}, {\bf y}) = \mathop{\rm max}_{i \in [1,N]}\limits |x_i - y_i|$

KL: A symmetrized Kullback-Leibler divergence: ${\rm KL}({\bf x}, {\bf y}) = \sum_{i=1}^N x_i \log(x_i/y_i) + y_i \log(y_i/x_i)$

CD: The "cosine" distance: ${\rm CD}({\bf x}, {\bf y}) = 1 - {\bf x} {\boldsymbol \cdot} {\bf y} / (\|{\bf x}\| \|{\bf y}\|)$

With a symmetric strictly positive definite weight matrix W this becomes: $\\ {\rm CD}({\bf x}, {\bf y}, \hbox{M=W}) = 1 - {\bf x} {\boldsymbol \cdot} \left( M {\bf y}\right) / (|\!|\!|{\bf x}|\!|\!| |\!|\!|{\bf y}|\!|\!|)$

Here: $|\!|\!| {\bf z} |\!|\!| = \sqrt{{\bf z} {\boldsymbol \cdot} \left( M {\bf z}\right)}$

JD: The Jaccard distance.

What is the best cluster number?

The function find_best_cluster attempts to find the best cluster number. To do this, it monitors the total variation as one increases the cluster number. The total variation goes down (generally) as we find (potentially locally) optimal solutions for each cluster number. If we pick a cluster number using only the total variation, we will miss the "natural cluster" number.

To avoid this, we adjust the total variation by a function that depends on the dimension of the space we are working in as well as the cluster number. The reasoning follows:

The idea is to look at the natural rate at which the total variation decreases with cluster number when there are no clusters. In this way we can adjust the total variation to take into account this "ambient" decay.

To do this, we start by assuming that the data is uniformly distributed in our domain (with respect to the metric used) when given the data: k clusters; m points; the domain in n dimensions. We assume that the k clusters have the same number of points and fill a sphere of radius, R. This means that $R^n \approx k \, r_k^n$.

Solving for $r_k$ we have ${r_k=R\\\left(\\\frac{1}{k}\\\right)^{\\\frac{1}{n}}}$. The total variation of k clusters is then roughly: ${k \\\, r_k\\\left(\\\frac{m}{k}\\\right)}$. This becomes: $\\\frac{m R}{k^{\\\frac{1}{n}}}$. Thus, even in the absence of any true clusters, the total variation decays like $k^{\\\frac{1}{n}}$.

The function find_best_cluster compares the total variation of cluster numbers in a range. It chooses the cluster number, k, with the largest relative rate of decrease (with respect to cluster size) in adjusted total variation. The adjusted variation modifies the total variation for each k by the multiplicative factor $k^{\\\frac{1}{n}}$. The variation is further adjusted by the fraction of unused cluster centroids. Finally, before computing the relative rate of variation decrease, the series is further adjusted to be monotonically non-increasing.

NOTE: This analysis may not be as useful if the "natural" clusters (or a substantial subset) lie in some lower dimensional hyperplane in the ambient space.

Cluster Functions

Cluster.kmeans_cluster — Method

kmeans_cluster(X, k=3[; dmetric, threshold, W, N, seed])

Groups a set of points into k clusters based on the distance metric, dmetric.

Type Constraints

T <: Real

F <: Function

Arguments

X::Matrix{T} : (n,m) Matrix representing m points of dimension n.

k::Int=3 : The number of clusters to form.

Keyword Arguments

dmetric::F=L2 : The distance metric to use.

threshold::Float=1.0e-2 : The relative error improvement threshold (using total variation)

W::Union{Nothing, AbstractMatrix{T}}=nothing : Optional (nxn) weight matrix for metric.

N::Int=1000 : The maximum number of iterations to try.

seed::Int=0 : If value > 0, create a random number generator to use for initial clustering.

Input Contract

$W = {\rm nothing} ∨ \left( ({\rm typeof}(W) = {\rm Matrix}\{T\}) ∧ W \in {\boldsymbol S}_{++}^{n} \right)$

$1 \le k \le m$

N > 0

threshold > 0.0

dmetric <: Function

Return

A Tuple:

Dict{Int, Int} : Mapping of points (n-vectors) indices to centroid indices.

Matrix{T} : (nxk) Matrix representing k centroids of n-vectors.

Float64 : The total variation between points and their centroids (using dmetric).

Vector{Int} : Unused centroids (by index).

Int : The number of iterations to use for the algorithm to complete.

Bool : Did algorithm converge.

Cluster.find_best_info_for_ks — Method

find_best_info_for_ks(X, kRng[; dmetric=L2, threshold=1.0e-3, W, N=1000, num_trials=100, seed=1])

Groups a set of m points (n-vectors) as an (nxm) matrix, X, into k clusters where k is in the range, kRng. The groupings are determined based on the distance metric, dmetric.

Type Constraints

T <: Real

F <: Function

Arguments

X::Matrix{T} : (n,m) Matrix representing m points of dimension n.

kRng::UnitRange{Int} : The number of clusters to form.

Keyword Arguments

dmetric::F=L2 : The distance metric to use.

threshold::Float=1.0e-2: The relative error improvement threshold (using total variation)

W::Union{Nothing, AbstractMatrix{T}}=nothing : Optional Weight matrix for metric.

N::Int=1000 : The maximum number of kmeans_clustering iterations to try for each cluster number.

num_trials::Int=300 : The number of times to run kmeans_clustering for a given cluster number.

seed::Int=1 : The random seed to use. (Used by kmeans_cluster to do initial clustering.)

Input Contract

$W = {\rm nothing} ∨ \left( ({\rm typeof}(W) = {\rm Matrix}\{T\}) ∧ W \in {\boldsymbol S}_{++}^{n} \right)$

$N > 0$

$∀ i \in {\rm kRng}, i \ge 1$

threshold > 0.0

Return

A Tuple with entries:

OrderedDict{Int, Float} : 1:k -> The Total Variation for each cluster number.

OrderedDict{Int, Vector{Int}} : 1:k -> Mapping of index of points (n-vectors in X) to centroid indices.

OrderedDict{Int, Matrix{T}} : 1:k -> (nxk) Matrix representing k n-vector centroids.

OrderedDict{Int, Vector{In64}} : 1:k -> Vector of unused centroids by index.

Cluster.find_best_cluster — Method

find_best_cluster(X, kRng[; dmetric=L2, threshold=1.0e-3, W, N=1000, num_trials=100, seed=1, verbose=false])

Groups a set of points into the "best" number of clusters based on the distance metric, dmetric. It does this by examining the total variation between the points and the centroids for groups of k where k is in the range, kRng.

NOTE: If the value k was determined to be the best cluster number but some of the centroids were not used, then the value of k will be set to the number of centroids that are used and the centroids that were not used will be removed. In this case it may be that the returned value of k is less that any value in the cluster range, kRng.

Type Constraints

T <: Real

F <: Function

Arguments

X::Matrix{T} : (n,m) Matrix representing m points of dimension n.

kRng::UnitRange{Int} : The range of potential cluster values to try.

Keyword Arguments

dmetric::F=L2 : The distance metric to use.

threshold::Float=1.0e-2: The relative error improvement threshold (using total variation)

W::Union{Nothing, AbstractMatrix{T}}=nothing : Optional Weight matrix for metric.

N::Int=1000 : The maximum number of kmeans_clustering iterations to try for each cluster number.

num_trials::Int=300 : The number of times to run kmeans_clustering for a given cluster number.

seed::Int=1 : The random seed to use. (Used by kmeans_cluster to do initial clustering.)

verbose::Bool=false : If true, print diagnostic information.

Input Contract

$W = {\rm nothing} ∨ \left( ({\rm typeof}(W) = {\rm Matrix}\{T\}) ∧ W \in {\boldsymbol S}_{++}^{n} \right)$

N > 0

$∀ i \in {\rm kRng}, i \ge 1$

threshold > 0.0

Return

A Tuple:

Int : The "best" cluster number, k.

Dict{Int, Int}: Mapping of points (n-vectors) indices to centroid indices.

Matrix{T} : Cluster centroids, represented as an (n,k) matrix.

Float64 : The total variation between points and their centroids (using dmetric).

Metric Functions

Cluster.Metrics.L2 — Method

L2(x,y[; M=nothing])

Computes the $L_2$ distance between two vectors. One of the features that may be different from other packages is the use of weighted metrics in some instances.

Type Constraints

T <: Real

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

Keyword Arguments

C::Union{Nothing, Matrix{T} : Optional Weight matrix.

Input Contract (Low level function – Input contract not checked)

$|{\bf x}| = |{\bf y}|$

$M = {\rm nothing} \vee \left( ({\rm typeof}(M) = {\rm Matrix}\{T\}) \wedge M \in {\boldsymbol S}_{++}^{|{\bf x}|} \right)$

Return

$L_2$ (optionally weighted) distance measure between the two vectors.

Cluster.Metrics.LP — Method

LP(x,y,p)

Computes the $L_p$ distance between two vectors.

Type Constraints

T <: Real

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

p::Int : The power of the norm.

Input Contract (Low level function – Input contract not checked)

$|{\bf x}| = |{\bf y}|$

p > 0

Return

$L_p$ distance measure between the two vectors.

Cluster.Metrics.LI — Method

LI(x,y)

Computes the $L_\infty$ distance between two vectors.

Type Constraints

T <: Real

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

Input Contract (Low level function – Input contract not checked)

$|{\bf x}| = |{\bf y}|$

Return

$L_\infty$ distance measure between the two vectors.

Cluster.Metrics.KL — Method

KL(x,y)

Computes the $Kullback-Leibler$ distance between two vectors.

Type Constraints

T <: Real

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

Input Contract (Low level function – Input contract not checked)

Let $N = |{\bf x}|$.

$|{\bf x}| = |{\bf y}|$

$\forall i \in [1, N]: x_i \ge 0$

$\forall i \in [1, N]: y_i \ge 0$

$\sum_{i=1}^N x_i = 1$

$\sum_{i=1}^N y_i = 1$

Return

KL distance measure between the two vectors.

Cluster.Metrics.CD — Method

CD(x,y[; M=nothing])

Computes the "cosine" distance between two vectors.

Type Constraints

T <: Real

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

Keyword Arguments

M::Union{Nothing, Matrix{T} : Optional Weight matrix.

Input Contract (Low level function – Input contract not checked)

$|{\bf x}| = |{\bf y}|$

$M = {\rm nothing} \vee \left( ({\rm typeof}(M) = {\rm Matrix}\{T\}) \wedge M \in {\boldsymbol S}_{++}^{|{\bf x}|} \right)$

Return

Cosine distance measure between the two vectors.

Cluster.Metrics.JD — Method

JD(x,y)

Computes the Jaccard metric between two vectors of a "discrete" type. For instance, the vectors could be integers; however, they can also be of non-numeric type. The metric can also be used with floating point values but, in that case, it may be more useful to round/truncate to a particular "block" size.

If both x and y are vectors of zero length, a distance of $0$ is returned.

Arguments

x::AbstractVector{T} : A numeric vector.

y::AbstractVector{T} : A numeric vector.

Return

Jaccard distance measure between the two vectors.

Fit Metric Functions

Cluster.Metrics.raw_confusion_matrix — Method

raw_confusion_matrix(act, pred)

Computes a confusion matrix of the discrete variables, act and pred. There are no row or column labels for this matrix.

Type Constraints:

Expects types, A and P to have discrete values.

Arguments

act ::AbstractVector{A} : A vector of the actual target values.

pred::AbstractVector{P} : A vector of the predicted target values.

Input Contract:

|act| = |pred|

Return

Tuple{Vector{A}, Vector{P}, Matrix{Int}}: A 3-tuple consisting of:

Vector of unique values of act. (Sorted from lowest to highest, otherwise the order returned from the function unique.)

Vector of unique values of pred. (Sorted from lowest to highest, otherwise the order returned from the function unique.)

A matrix of counts for all pairings of discrete values of act with pred.

Cluster.Metrics.confusion_matrix — Method

confusion_matrix(act, pred)

Computes the confusion matrix of the discrete variables, act and pred.

Type Constraints:

Expects types, A and P to have discrete values.

Arguments

act ::AbstractVector{A} : A vector of the actual target values.

pred::AbstractVector{P} : A prediction vector for the target.

Input Contract:

|act| = |pred|

Return

Matrix{Any}:

The raw confusion matrix augmented by a column on the left listing all actual values (in sorted order if sortable) and augmented on top with a row listing all predicted values (in sorted order if sortable).

Index

Cluster.Metrics.CD

Cluster.Metrics.JD

Cluster.Metrics.KL

Cluster.Metrics.L2

Cluster.Metrics.LI

Cluster.Metrics.LP

Cluster.Metrics.confusion_matrix

Cluster.Metrics.raw_confusion_matrix

Cluster.find_best_cluster

Cluster.find_best_info_for_ks

Cluster.kmeans_cluster